+++ kiosk/default.nix
@@ -15,15 +15,24 @@ python3Packages.buildPythonApplication rec {
       --replace "@system_version@" "${system_version}"
   '';
 
-  doCheck = false;
+  buildInputs = [
+    bashInteractive
+    makeWrapper
+  ];
 
-  nativeBuildInputs = [ qt5.wrapQtAppsHook mypy ];
+  nativeBuildInputs = [
+    mypy
+    qt6.wrapQtAppsHook
+  ];
 
   propagatedBuildInputs = with python3Packages; [
     dbus-python
     pygobject3
-    pyqtwebengine
+    pyqt6-webengine
     pytest
+    qt6.full
+    qt6.qtbase
+    qt6.qtbase
     requests
     types-requests
   ];
@@ -32,15 +41,14 @@ python3Packages.buildPythonApplication rec {
     cp -r images/ $out/images
   '';
 
-  dontWrapQtApps = true;
-  makeWrapperArgs = [ "\${qtWrapperArgs[@]}" ];
-
   shellHook = ''
     # Give access to kiosk_browser module
     export PYTHONPATH=./:$PYTHONPATH
 
-    # Give access to Qt platform plugin "xcb" in nix-shell
-    export QT_QPA_PLATFORM_PLUGIN_PATH="${qt5.qtbase.bin}/lib/qt-${qt5.qtbase.version}/plugins";
+    # Setup Qt environment
+    export QT_QPA_PLATFORM=wayland
+    bashdir=$(mktemp -d)
+    makeWrapper "$(type -p bash)" "$bashdir/bash" "''${qtWrapperArgs[@]}"
+    exec "$bashdir/bash"
   '';
-
 }