+++ controller/Changelog.md
@@ -11,6 +11,7 @@
 - os: Set noexec for volatile root and persistent storage mounts
 - os: Restrict remote maintenance to the ZeroTier network
 - os: Limit permitted SSH modes and forwarding options
+- kiosk: Migrate to Qt6
 
 ## Removed